further cpp03 changes

+++ groups/bsl/bslfmt/bslfmt_formatarg.h
@@ -410,7 +410,7 @@ class basic_format_args {
 
     // FRIENDS
     template <class t_INNER_CONTEXT>
-    friend int Format_FormatArgsSize(
+    friend size_t Format_FormatArgsSize(
             const basic_format_args<t_INNER_CONTEXT>& args);
 };
 
@@ -419,7 +419,7 @@ typedef basic_format_args<format_context> format_args;
 typedef basic_format_args<wformat_context> wformat_args;
 
 template<class t_CONTEXT>
-int Format_FormatArgsSize(const basic_format_args<t_CONTEXT>& args)
+size_t Format_FormatArgsSize(const basic_format_args<t_CONTEXT>& args)
     // This component-private function returns the result of calling 'size()'
     // on the specified 'args' parameter. This is to permit access to the
     // private 'size' accessor of 'basic_format_args' without requiring long
